Wrapper function to load arbitrary VCF files.

Wrapper function to load arbitrary VCF files.

load_vcf_file(f, ..., verbose = TRUE, annot = TRUE)

Arguments

f
File name

...
Other optional arguments.

verbose
Verbosity flag (Default: TRUE)

annot
Flag indicating if variant annotations should be added (Default: TRUE)

Value

A 'CollapsedVCF' object.

Examples

load_vcf_file(system.file("extdata", "platypus.vcf.bgz", package = "THmisc"))

#> --------------- Loading a VCF file ---------------
#> 
#>   File infos:
#>    - Name: 'platypus.vcf.bgz'
#>    - Created: 2021-09-02 17:04:18 
#>    - Modified: 2021-09-02 17:04:16 
#> 
#>   Identifying type... 
#>    => Type is 'Platpyus'.
#> 
#>   VCF object contains:
#>    => N=2 variants.
#>    => Of these 2/2 (100%) marked passing.
#>    => On 1 chromosomes
#> 
#>   Guessing reference of a vcf file:

#> 
#> Attaching package: ‘BiocGenerics’

#> The following objects are masked from ‘package:parallel’:
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB

#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs

#> The following objects are masked from ‘package:base’:
#> 
#>     anyDuplicated, append, as.data.frame, basename, cbind, colnames,
#>     dirname, do.call, duplicated, eval, evalq, Filter, Find, get, grep,
#>     grepl, intersect, is.unsorted, lapply, Map, mapply, match, mget,
#>     order, paste, pmax, pmax.int, pmin, pmin.int, Position, rank,
#>     rbind, Reduce, rownames, sapply, setdiff, sort, table, tapply,
#>     union, unique, unsplit, which.max, which.min

#> 
#> Attaching package: ‘S4Vectors’

#> The following objects are masked from ‘package:base’:
#> 
#>     expand.grid, I, unname

#> 
#> Attaching package: ‘Biostrings’

#> The following object is masked from ‘package:base’:
#> 
#>     strsplit

#> 
#> Attaching package: ‘BSgenome.Hsapiens.UCSC.hg19’

#> The following object is masked from ‘package:BSgenome.Hsapiens.UCSC.hg38’:
#> 
#>     Hsapiens

#> 
#> Attaching package: ‘BSgenome.Hsapiens.UCSC.hg18’

#> The following object is masked from ‘package:BSgenome.Hsapiens.UCSC.hg19’:
#> 
#>     Hsapiens

#> The following object is masked from ‘package:BSgenome.Hsapiens.UCSC.hg38’:
#> 
#>     Hsapiens

#>    => Testing 3/3 available options.
#>    => Found 2/3 perfectly matching.
#>    => Reference is 'BSgenome.Hsapiens.UCSC.hg38'.

#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.

#> 
#>   Generating annotations...
#>   
#>  -------------------------------------------------- 

#> class: CollapsedVCF 
#> dim: 2 2 
#> rowRanges(vcf):
#>   GRanges with 6 metadata columns: paramRangeID, ANNOTATION, REF, ALT, QUAL, FILTER
#> info(vcf):
#>   DataFrame with 25 columns: FR, MMLQ, TCR, HP, WE, Source, FS, WS, PP, TR, ...
#> info(header(vcf)):
#>                   Number Type    Description                                   
#>    FR             .      Float   Estimated population frequency of variant     
#>    MMLQ           1      Float   Median minimum base quality for bases aroun...
#>    TCR            1      Integer Total reverse strand coverage at this locus   
#>    HP             1      Integer Homopolymer run length around variant locus   
#>    WE             1      Integer End position of calling window                
#>    Source         .      String  Was this variant suggested by Playtypus, As...
#>    FS             .      Float   Fisher's exact test for strand bias (Phred ...
#>    WS             1      Integer Starting position of calling window           
#>    PP             .      Float   Posterior probability (phred scaled) that t...
#>    TR             .      Integer Total number of reads containing this variant 
#>    NF             .      Integer Total number of forward reads containing th...
#>    TCF            1      Integer Total forward strand coverage at this locus   
#>    NR             .      Integer Total number of reverse reads containing th...
#>    TC             1      Integer Total coverage at this locus                  
#>    END            .      Integer End position of reference call block          
#>    MGOF           .      Integer Worst goodness-of-fit value reported across...
#>    SbPval         .      Float   Binomial P-value for strand bias test         
#>    START          .      Integer Start position of reference call block        
#>    ReadPosRankSum .      Float   Mann-Whitney Rank sum test for difference b...
#>    MQ             .      Float   Root mean square of mapping qualities of re...
#>    QD             1      Float   Variant-quality/read-depth for this variant   
#>    SC             1      String  Genomic sequence 10 bases either side of va...
#>    BRF            1      Float   Fraction of reads around this variant that ...
#>    HapScore       .      Integer Haplotype score measuring the number of hap...
#>    Size           .      Integer Size of reference call block                  
#> geno(vcf):
#>   List of length 7: GT, GQ, GOF, NR, GL, NV, VAF
#> geno(header(vcf)):
#>        Number Type    Description                                              
#>    GT  1      String  Unphased genotypes                                       
#>    GQ  .      Integer Genotype quality as phred score                          
#>    GOF .      Float   Goodness of fit value                                    
#>    NR  .      Integer Number of reads covering variant location in this sample 
#>    GL  .      Float   Genotype log10-likelihoods for AA,AB and BB genotypes,...
#>    NV  .      Integer Number of reads containing variant in this sample        
#>    VAF .      Float   Variant allele frequency